requirements.txt
flask==3.0.3
PyPDF2==3.0.1
python-pptx==1.0.2
python-docx==1.1.2
anthropic>=0.40.0
